requirements.txt
Flask==1.0.2
Jinja2==2.10
psycopg2
python-dotenv
loguru
Flask-SQLAlchemy===2.4.1
Flask-Migrate==2.5.2